Better `isClass` function, closes #28
