+++ lib/loadModules.js
@@ -1,5 +1,5 @@
 const camelCase = require('camel-case')
-const isClass = require('is-class')
+const isClass = require('./isClass')
 const isString = require('is-string')
 const isFunction = require('./isFunction')
 const Lifetime = require('./Lifetime')
