+++ package.json
@@ -54,7 +54,6 @@
   "dependencies": {
     "camel-case": "^3.0.0",
     "glob": "^7.1.2",
-    "is-class": "0.0.4",
     "is-plain-object": "^2.0.4",
     "is-string": "^1.0.4"
   }
